+++ scripts/update-abi.js
@@ -1,27 +1,42 @@
+#!/usr/bin/env node
+
 const { execSync } = require('child_process')
-const { writeFileSync } = require('fs')
+const { writeFile } = require('fs/promises')
 const { resolve } = require('path')
 const { fetch } = require('undici')
 
 ;(async () => {
-  const [{ abi: abi1 }, { abi: abi2 }] = await Promise.all([
-    fetch(
-      'https://raw.githubusercontent.com/Crossbell-Box/Crossbell-Contracts/main/build-info/Web3Entry.json',
-    ).then((res) => res.json()),
-    fetch(
-      'https://raw.githubusercontent.com/Crossbell-Box/Crossbell-Contracts/main/build-info/Events.json',
-    ).then((res) => res.json()),
-  ])
+  const [{ abi: abi1 }, { abi: abi2 }, { abi: periphery_abi }] =
+    await Promise.all([
+      fetch(
+        'https://raw.githubusercontent.com/Crossbell-Box/Crossbell-Contracts/main/build-info/Web3Entry.json',
+      ).then((res) => res.json()),
+      fetch(
+        'https://raw.githubusercontent.com/Crossbell-Box/Crossbell-Contracts/main/build-info/Events.json',
+      ).then((res) => res.json()),
+      fetch(
+        'https://raw.githubusercontent.com/Crossbell-Box/Crossbell-Contracts/main/build-info/Periphery.json',
+      ).then((res) => res.json()),
+    ])
 
   const abi = [...abi1, ...abi2]
 
-  writeFileSync(
-    resolve(__dirname, '../src/contract/abi/abi.json'),
-    JSON.stringify(abi, null, 2),
+  const entryDir = resolve(__dirname, '../src/contract/abis/entry')
+  const peripheryDir = resolve(__dirname, '../src/contract/abis/periphery')
+
+  await writeFile(`${entryDir}/abi.json`, JSON.stringify(abi, null, 2))
+
+  await writeFile(
+    `${peripheryDir}/abi.json`,
+    JSON.stringify(periphery_abi, null, 2),
   )
 
   execSync(
-    'npx typechain --target ethers-v5 --out-dir ./src/contract/abi/types ./src/contract/abi/abi.json',
+    `npx typechain --target ethers-v5 --out-dir ${entryDir}/types ${entryDir}/abi.json`,
+  )
+
+  execSync(
+    `npx typechain --target ethers-v5 --out-dir ${peripheryDir}/types ${peripheryDir}/abi.json`,
   )
 
   console.log('done')