+++ src/index.ts
@@ -1,9 +1,8 @@
 import { Network } from './network'
 import { Contract } from './contract'
 import { Indexer } from './indexer'
+import { Ipfs } from './ipfs'
 
-export default { Network, Contract, Indexer }
-export { Network, Contract, Indexer }
-
-export * from './indexer/types'
-export * from './contract/types'
+export default { Network, Contract, Indexer, Ipfs }
+export { Network, Contract, Indexer, Ipfs }
+export * from './types'